+++ packages/react-native-nitro-http/cpp/ServerCore.hpp
@@ -54,7 +54,7 @@ class ServerCore {
 public:
   ServerCore() = default;
 
-  void listen(double port);
+  void listen(int port);
   void stop();
 
   OfferCallback _offerCb;
